+++ source/components/CustomScreenMovementManager.h
@@ -62,8 +62,8 @@
         /* update axis flag */																			\
         VBVec3DFlag positionFlag;																		\
         /* pointer to function that implements the required focusing */									\
-        bool (*focusFunction)(CustomScreenMovementManager, u8, u8);                                     \
-        bool (*previuosFocusFunction)(CustomScreenMovementManager, u8, u8);                             \
+        bool (*focusFunction)(CustomScreenMovementManager, u32, u32);                                   \
+        bool (*previuosFocusFunction)(CustomScreenMovementManager, u32, u32);                           \
 
 // declare a CustomScreenMovementManager
 __CLASS(CustomScreenMovementManager);
